services/document-worker/requirements.txt
httpx>=0.26.0
redis>=5.0.0
pypdf>=3.17.0
python-docx>=1.1.0
Pillow>=10.0.0
pytesseract>=0.3.10
